Update README.md
This project implements a state-of-the-art facial recognition model utilizing deep learning techniques. The goal is to build a robust system that can accurately identify and verify individuals from images. With the rapid advancements in machine learning and image processing, facial recognition has become an essential component in various applications, such as security systems, access control, and social media tagging.